+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Layouts
 
+- Add `AnalyticsShell` component. [#307](https://github.com/mapbox/dr-ui/pull/307)
+
 ## 0.29.2
 
 - Add aria-label to `BackToTopButton`. [#290](https://github.com/mapbox/dr-ui/pull/290)