+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `AnalyticsShell` component. [#307](https://github.com/mapbox/dr-ui/pull/307)
 - Add Batfish helpers: navigation and topics.
+- Update PageLayout to accept layouts.
 
 ## Main