+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Main
+## 0.30.0
 
 - Add `ErrorBoundary` component. [#272](https://github.com/mapbox/dr-ui/pull/272)
 - Pin the version numbers for all search-ui dependencies. [#321](https://github.com/mapbox/dr-ui/pull/321)